spike out using build flags
